move traits to utils
