+++ utils/forest_utils/src/db/mod.rs
@@ -1,6 +1,8 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+pub mod file_backed_obj;
+
 use cid::{
     multihash::{Code, MultihashDigest},
     Cid,
